requirements_devel.txt
#
# Requirements for wger during development
#

# Production packages
-r requirements.txt

# Development packages
pep8
django-debug-toolbar
